2nd commit (small fixes + test services)

+++ src/main/java/com/example/E_CommerceSCD/DTOs/ProductCreateUpdateDTO.java
@@ -14,5 +14,5 @@ public class ProductCreateUpdateDTO {
     private String description;
     private Double price;
     private int stockQuantity;
-    private int categoryId;
+    private Long categoryId;
 }
